+++ chapter-05/squashfs-root/usr/sbin/httpd.dec/sub_159DC@159DC.c
@@ -1,6 +1,6 @@
 size_t sub_159DC()
 {
   sub_13954("\nwlnoise = [");
-  foreach_wif(0, 0, sub_14980);
+  foreach_wif(0, 0, (int)sub_14980);
   return sub_13954(" ];\n");
 }
