+++ chapter-05/squashfs-root/usr/sbin/httpd.dec/sub_15A10@15A10.c
@@ -2,7 +2,7 @@ size_t sub_15A10()
 {
   char *v0; // r0
 
-  if ( foreach_wif(1, 0, sub_14534) )
+  if ( foreach_wif(1, 0, (int)sub_14534) )
     v0 = (char *)"0";
   else
     v0 = (char *)"1";
