+++ chapter-05/squashfs-root/usr/sbin/httpd.dec/sub_15A4C@15A4C.c
@@ -7,6 +7,6 @@ size_t __fastcall sub_15A4C(int a1, const char **a2)
   else
     v2 = 0;
   sub_13954("\nwlstats = [");
-  foreach_wif(v2, 0, sub_1461C);
+  foreach_wif(v2, 0, (int)sub_1461C);
   return sub_13954("];\n");
 }
