+++ chapter-05/squashfs-root/usr/sbin/httpd.dec/sub_15D2C@15D2C.c
@@ -7,6 +7,6 @@ size_t __fastcall sub_15D2C(int a1, const char **a2)
   else
     v2 = 0;
   sub_13954("\nwl_ifaces = [");
-  foreach_wif(v2, 0, sub_1416C);
+  foreach_wif(v2, 0, (int)sub_1416C);
   return sub_13954("];\n");
 }
