+++ chapter-05/squashfs-root/usr/sbin/httpd.str/266CF_wldev = [/sub_16798@16798.c
@@ -21,7 +21,7 @@ int sub_16798()
   sub_159DC();
   sub_13954("wldev = [");
   v16 = 32;
-  foreach_wif(1, &v16, sub_162AC);
+  foreach_wif(1, (int)&v16, (int)sub_162AC);
   sub_13954("];\n");
   v14 = "wan_ifname,wan2_ifname,wan3_ifname,wan4_ifname,wan_iface,wan2_iface,wan3_iface,wan4_iface,wan_proto,wan2_proto,w"
         "an3_proto,wan4_proto,wan_ifnameX,wan2_ifnameX,wan3_ifnameX,wan4_ifnameX,wan_ifnames,wan2_ifnames,wan3_ifnames,wa"
