+++ chapter-05/squashfs-root/usr/sbin/httpd.str/2715D_%s_modem_dev/sub_1923C@1923C.c
@@ -5,17 +5,17 @@ const char *__fastcall sub_1923C(const char *a1)
   char v4[48]; // [sp+0h] [bp-30h] BYREF
 
   snprintf(v4, 0x20u, "%s_proto", a1);
-  if ( is_value_in_nvram_1(v4, "ppp3g") )
+  if ( is_value_in_nvram_1((int)v4, "ppp3g") )
     goto LABEL_6;
-  if ( !is_value_in_nvram_1(v4, "lte") )
+  if ( !is_value_in_nvram_1((int)v4, "lte") )
   {
     v3 = "\nwwansms_error = 'WAN is not WWAN!'";
     goto LABEL_11;
   }
   snprintf(v4, 0x20u, "%s_modem_type", a1);
-  if ( !is_value_in_nvram_1(v4, "non-hilink")
-    && !is_value_in_nvram_1(v4, "huawei-non-hilink")
-    && !is_value_in_nvram_1(v4, "hw-ether") )
+  if ( !is_value_in_nvram_1((int)v4, "non-hilink")
+    && !is_value_in_nvram_1((int)v4, "huawei-non-hilink")
+    && !is_value_in_nvram_1((int)v4, "hw-ether") )
   {
     v3 = "\nwwansms_error = 'WWAN is not supported!'";
 LABEL_11:
